backend/requirements.txt
alembic==1.13.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.112.2
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.3
httpx==0.27.0
python-multipart==0.0.12
ruff==0.6.9
mypy==1.11.2
watchdog==4.0.1
slowapi==0.1.9
email-validator==2.1.1
